add genesis json unmarshal test
Test that the legacy solo machines can be successfully unmarshalled.
This requires registering an implementation for the legacy solo machine. An implemenation which panics has been added.
This implementation should only be registered against a clientCtx during a migrate cli cmd. The implementation is only briefly used in order to decode the previous solo machine set in genesis.